Fix broken tests.
