+++ package.json
@@ -10,7 +10,7 @@
     "@fortawesome/free-solid-svg-icons": "5.15.3",
     "@fortawesome/react-fontawesome": "0.1.14",
     "ant-design-pro": "2.3.2",
-    "antd": "4.16.3",
+    "antd": "4.16.5",
     "elementary-editor": "0.9.15",
     "global": "^4.4.0",
     "react": "17.0.2",
@@ -67,7 +67,7 @@
     "@babel/plugin-transform-modules-commonjs": "7.14.5",
     "@babel/preset-es2015": "^7.0.0-beta.53",
     "@testing-library/jest-dom": "5.14.1",
-    "@testing-library/react": "11.2.7",
+    "@testing-library/react": "12.0.0",
     "antd-dayjs-webpack-plugin": "1.0.6",
     "babel-plugin-import": "1.13.3",
     "customize-cra": "^1.0.0",
