--- src/Drafts/test/data.js
@@ -1,23 +1,0 @@
-const storyData = {
-  id: 35,
-  date: "2019-07-21 20:57:24.238995",
-  title: 'fasddfsadsfa',
-  category: {
-    id: 8,
-    name: 'Category X'
-  },
-  content: '{"blocks": [{"key": "bhpma", "text": "Category X", "type": "unstyled", "depth": 0, "inlineStyleRanges": [], "entityRanges": [], "data": {}}], "entityMap": {}}',
-  comments: [],
-  tags: ['Ars'],
-  isFetching: false,
-  error: false,
-  author: {
-    id: 13,
-    name: '',
-    avatar: 'admin_d947a61b-a4a3-4858-858e-61d6b58dac41..png',
-    username: 'admin',
-    signature: 'Base signature.'
-  }
-};
-
-export default storyData;
